Updating for D01

+++ setup/reset.sh
@@ -1,5 +1,10 @@
 #!/bin/bash
 
+if [[ $EUID -ne 0 ]]; then
+   echo " [!]This script must be run as root" 1>&2
+   exit 1
+fi
+
 IFS='/' read -a array <<< pwd
 
 if [[ "$(pwd)" != *setup ]]
@@ -8,15 +13,29 @@ then
 fi
 
 # reset the database
-rm ../data/empire.db
+if [ -e ../data/empire.db ]
+then
+	rm ../data/empire.db
+fi
+
 ./setup_database.py
 cd ..
 
 # remove the debug file if it exists
-rm empire.debug
+if [ -e empire.debug ]
+then
+	rm empire.debug
+fi
 
 # remove the download folders
-rm -rf ./downloads/
+if [ -d ./downloads/ ]
+then
+	rm -rf ./downloads/
+fi
 
-# start up Empire
-./empire
+# start up Empire if not in docker otherwise return
+if [ -f /.dockerenv ]; then
+    echo " [*] Empire reset complete returning back to Docker"
+else
+    ./empire
+fi
